refactor: /notice 목록, 뷰어 리팩터링
