+++ actions/sessionActions.ts
@@ -11,7 +11,7 @@ import { COOKIE_SESSION_ID } from '@/constants/network';
 export const getMockAuth = async () => {
   // TODO: getRequest 함수로 구현
   // 지금은 /v1이 segment에 없어서 임시로 fetch를 사용해 직접 구현
-  const resp = await fetch(`https://cse-dev-waffle.bacchus.io/api/mock-login`, {
+  const resp = await fetch(`https://cse-dev-waffle.bacchus.io/api/v1/mock-login`, {
     method: 'GET',
   });
