+++ app/[locale]/community/notice/[id]/edit/EditNoticePageContent.tsx
