+++ app/[locale]/community/notice/[id]/edit/page.tsx
@@ -1,6 +1,6 @@
 import { getNoticePostDetail } from '@/apis/notice';
 
-import EditNoticePageContent from '@/app/[locale]/community/notice/components/EditNoticePageContent';
+import EditNoticePageContent from '@/app/[locale]/community/notice/[id]/edit/EditNoticePageContent';
 
 interface EditNoticePageProps {
   params: { id: string };
