+++ app/[locale]/community/notice/[id]/page.tsx
@@ -1,6 +1,5 @@
 import { getNoticePostDetail } from '@/apis/notice';
 
-import AdjPostNav from '@/components/common/AdjPostNav';
 import Attachments from '@/components/common/Attachments';
 import { StraightNode } from '@/components/common/Nodes';
 import Tags from '@/components/common/Tags';
@@ -11,6 +10,7 @@ import PageLayout, {
   PAGE_PADDING_RIGHT_PX,
   PAGE_PADDING_TOP_PX,
 } from '@/components/layout/pageLayout/PageLayout';
+import AdjPostNav from '@/components/post/AdjPostNav';
 
 import { PostSearchQueryParams } from '@/types/post';
 
@@ -34,11 +34,11 @@ export default async function NoticePostPage({
   // ID가 잘못된 경우 예외 처리
   if (Number.isNaN(id)) return <InvalidIDFallback rawID={rawID} />;
 
-  const currPost = await getNoticePostDetail(id, searchParams);
+  const notice = await getNoticePostDetail(id, searchParams);
 
   return (
     <PageLayout titleType="big" titleMargin="mb-5" bodyStyle={{ padding: 0 }}>
-      <Header title={currPost.title} author={currPost.author} dateStr={currPost.createdAt} />
+      <Header {...notice} />
       <div
         className="bg-neutral-50 pt-9"
         style={{
@@ -47,11 +47,11 @@ export default async function NoticePostPage({
           paddingBottom: PAGE_PADDING_BOTTOM_PX,
         }}
       >
-        <Attachments files={currPost.attachments} />
-        <HTMLViewer htmlContent={currPost.description} className="mb-10 mt-4" />
+        <Attachments files={notice.attachments} />
+        <HTMLViewer htmlContent={notice.description} className="mb-10 mt-4" />
         <StraightNode />
-        <Tags tags={currPost.tags} margin="mt-3 ml-6" searchPath={noticePath} />
-        <AdjPostNav post={currPost} postType="notice" id={rawID} margin="mt-12" />
+        <Tags tags={notice.tags} margin="mt-3 ml-6" searchPath={noticePath} />
+        <AdjPostNav post={notice} postType="notice" id={rawID} margin="mt-12" />
       </div>
     </PageLayout>
   );
@@ -65,7 +65,15 @@ const InvalidIDFallback = ({ rawID }: { rawID: string }) => (
   </PageLayout>
 );
 
-const Header = ({ title, author, dateStr }: { title: string; author: string; dateStr: string }) => {
+const Header = ({
+  title,
+  author,
+  createdAt,
+}: {
+  title: string;
+  author: string;
+  createdAt: string;
+}) => {
   return (
     <div
       className="mb-9 flex flex-col gap-4"
@@ -74,7 +82,7 @@ const Header = ({ title, author, dateStr }: { title: string; author: string; dat
       <h2 className="text-[1.25rem] font-semibold">{title}</h2>
       <div className="flex gap-5 text-sm font-normal tracking-wide text-neutral-500">
         <p>글쓴이: {author}</p>
-        <p>작성 시각: {formatPostDateStr(new Date(dateStr))}</p>
+        <p>작성 시각: {formatPostDateStr(createdAt)}</p>
       </div>
     </div>
   );
